deleted login page, now u can log in using the nav button directly
