--- src/pages/Login.tsx
@@ -1,9 +1,0 @@
-import LoginForm from "../components/LoginForm";
-
-export default function Login() {
-  return (
-    <div>
-      <LoginForm />
-    </div>
-  );
-}
